Sql schema models

+++ server/models/Moon.js
@@ -20,7 +20,7 @@ Moon.init(
 			allowNull: false,
 		},
 		size: {
-			type: DataTypes.BIGINT,
+			type: DataTypes.FLOAT(10, 1),
 			allowNull: false,
 		},
 		planet_id: {
